Refactor of names
