+++ src/ts/App/Models/kingModel.ts
@@ -1,15 +1,15 @@
-import { IChessman, Color, ChessmanType, ListOfVectors } from './chessmanInterface';
+import { IFigure, Color, FigureType, ListOfVectors } from './figureInterface';
 
-export default class KingModel implements IChessman {
+export default class KingModel implements IFigure {
     readonly color: Color;
-    readonly name: ChessmanType;
+    readonly name: FigureType;
     is_moved: boolean;
     move_vectors: ListOfVectors;
     attack_vectors: ListOfVectors;
 
     constructor(color: Color) {
         this.color = color;
-        this.name = ChessmanType.King;
+        this.name = FigureType.King;
         this.move_vectors = [
             [[-1, -1]],
             [[-1, 0]],
