+++ src/ts/App/Models/knightModel.ts
@@ -1,15 +1,15 @@
-import { IChessman, Color, ChessmanType, ListOfVectors } from './chessmanInterface';
+import { IFigure, Color, FigureType, ListOfVectors } from './figureInterface';
 
-export default class KnightModel implements IChessman {
+export default class KnightModel implements IFigure {
     readonly color: Color;
-    readonly name: ChessmanType;
+    readonly name: FigureType;
     is_moved: boolean;
     move_vectors: ListOfVectors;
     attack_vectors: ListOfVectors;
 
     constructor(color: Color) {
         this.color = color;
-        this.name = ChessmanType.Knight;
+        this.name = FigureType.Knight;
         this.move_vectors = [
             [[-2, -1]],
             [[-2, 1]],
