+++ src/ts/App/Models/queenModel.ts
@@ -1,15 +1,15 @@
-import { IChessman, Color, ChessmanType, ListOfVectors } from './chessmanInterface';
+import { IFigure, Color, FigureType, ListOfVectors } from './figureInterface';
 
-export default class QueenModel implements IChessman {
+export default class QueenModel implements IFigure {
     readonly color: Color;
-    readonly name: ChessmanType;
+    readonly name: FigureType;
     is_moved: boolean;
     move_vectors: ListOfVectors;
     attack_vectors: ListOfVectors;
 
     constructor(color: Color) {
         this.color = color;
-        this.name = ChessmanType.Queen;
+        this.name = FigureType.Queen;
         this.move_vectors = [
             [[1, 1], [2, 2], [3, 3], [4, 4], [5, 5], [6, 6], [7, 7]], //right-bottom
             [[-1, 1], [-2, 2], [-3, 3], [-4, 4], [-5, 5], [-6, 6], [-7, 7]], //right-top
