+++ src/ts/App/Models/towerModel.ts
@@ -1,15 +1,15 @@
-import { IChessman, Color, ChessmanType, ListOfVectors } from './chessmanInterface';
+import { IFigure, Color, FigureType, ListOfVectors } from './figureInterface';
 
-export default class TowerModel implements IChessman {
+export default class TowerModel implements IFigure {
     readonly color: Color;
-    readonly name: ChessmanType;
+    readonly name: FigureType;
     is_moved: boolean;
     move_vectors: ListOfVectors;
     attack_vectors: ListOfVectors;
 
     constructor(color: Color) {
         this.color = color;
-        this.name = ChessmanType.Tower;
+        this.name = FigureType.Tower;
         this.move_vectors = [
             [[1, 0], [2, 0], [3, 0], [4, 0], [5, 0], [6, 0], [7, 0]], //top
             [[-1, 0], [-2, 0], [-3, 0], [-4, 0], [-5, 0], [-6, 0], [-7, 0]], //bottom
